Move Media related classes to own namespace
